protocols/posix: use dedicated messages for `EPOLL_CALL` requests
