+++ posix/subsystem/src/requests/common.hpp
@@ -137,6 +137,10 @@ struct HandleRequest {
 	operator()(managarm::posix::CloseRequest &&req, helix::BorrowedDescriptor conversation,
 			bragi::preamble preamble, std::shared_ptr<Process> self,
 			std::shared_ptr<Generation> generation);
+	async::result<std::expected<void, DispatchError>>
+	operator()(managarm::posix::EpollCallRequest &&req, helix::BorrowedDescriptor conversation,
+			bragi::preamble preamble, std::shared_ptr<Process> self,
+			std::shared_ptr<Generation> generation);
 
 	// From filesystem.cpp
 	async::result<std::expected<void, DispatchError>>
