+++ posix/subsystem/src/requests/fd.cpp
@@ -1,5 +1,8 @@
-#include "common.hpp"
 #include <fcntl.h>
+#include <sys/poll.h>
+
+#include "common.hpp"
+#include "../epoll.hpp"
 
 namespace requests {
 
@@ -152,6 +155,197 @@ HandleRequest::operator()(managarm::posix::CloseRequest &&req,
 	);
 	HEL_CHECK(sendResp.error());
 	logBragiReply(resp);
+	co_return {};
+}
+
+async::result<std::expected<void, DispatchError>>
+HandleRequest::operator()(managarm::posix::EpollCallRequest &&req,
+		helix::BorrowedDescriptor conversation, bragi::preamble preamble,
+		std::shared_ptr<Process> self, std::shared_ptr<Generation>) {
+	auto tailRes = co_await dispatchTail(req, conversation, preamble);
+	if(!tailRes) {
+		std::cout << "posix: Rejecting request due to decoding failure" << std::endl;
+		co_return std::unexpected(tailRes.error());
+	}
+
+	id = preamble.id();
+	logBragiRequest(req);
+
+	logRequest(logRequests, self, "EPOLL_CALL");
+
+	// Since file descriptors may appear multiple times in a poll() call,
+	// we need to de-duplicate them here.
+	struct PollEvents {
+		unsigned int events;
+		unsigned int revents;
+	};
+	std::unordered_map<int, PollEvents> fdsToEvents;
+
+	auto epfile = epoll::createFile();
+	assert(req.fds_size() == req.events_size());
+
+	auto timeout = req.timeout();
+
+	// First pass: build the deduplication map, merging events for duplicate FDs.
+	for(size_t i = 0; i < req.fds_size(); i++) {
+		// if fd is < 0, `events` shall be ignored and revents set to 0
+		if (req.fds(i) < 0) {
+			fdsToEvents.insert({req.fds(i), {0, 0}});
+			continue;
+		}
+
+		// Translate POLL events to EPOLL events.
+		if(req.events(i) & ~(POLLIN | POLLPRI | POLLOUT | POLLRDHUP | POLLERR | POLLHUP
+				| POLLNVAL | POLLWRNORM | POLLRDNORM)) {
+			std::cout << "\e[31mposix: Unexpected events for poll()\e[39m" << std::endl;
+			co_await sendErrorResponse<managarm::posix::EpollCallResponse>(conversation, managarm::posix::Errors::ILLEGAL_ARGUMENTS);
+			co_return {};
+		}
+
+		unsigned int mask = 0;
+		if(req.events(i) & POLLIN) mask |= EPOLLIN;
+		if(req.events(i) & POLLRDNORM) mask |= EPOLLIN;
+		if(req.events(i) & POLLOUT) mask |= EPOLLOUT;
+		if(req.events(i) & POLLWRNORM) mask |= EPOLLOUT;
+		if(req.events(i) & POLLPRI) mask |= EPOLLPRI;
+		if(req.events(i) & POLLRDHUP) mask |= EPOLLRDHUP;
+		if(req.events(i) & POLLERR) mask |= EPOLLERR;
+		if(req.events(i) & POLLHUP) mask |= EPOLLHUP;
+
+		auto [mapIt, inserted] = fdsToEvents.insert({req.fds(i), {mask, 0}});
+		if(!inserted)
+			mapIt->second.events |= mask; // Merge events for duplicate FDs.
+	}
+
+	// Second pass: add items to epoll and handle invalid FDs.
+	for(auto &[fd, pollEv] : fdsToEvents) {
+		if(fd < 0)
+			continue;
+
+		auto file = self->fileContext()->getFile(fd);
+		if(!file) {
+			// poll() is supposed to fail on a per-FD basis.
+			pollEv.revents = POLLNVAL;
+			timeout = 0;
+			continue;
+		}
+		auto locked = file->weakFile().lock();
+		assert(locked);
+
+		Error ret = epoll::addItem(epfile.get(), self.get(),
+			std::move(locked), fd, pollEv.events, fd);
+		assert(ret == Error::success);
+	}
+
+	struct epoll_event events[16] = {};
+	size_t k = 0;
+	bool interrupted = false;
+
+	auto cancelEvent = self->cancelEventRegistry().event(self->credentials(), req.cancellation_id());
+	if (!cancelEvent) {
+		std::println("posix: possibly duplicate cancellation ID registered");
+		co_await sendErrorResponse<managarm::posix::EpollCallResponse>(conversation, managarm::posix::Errors::INTERNAL_ERROR);
+		co_return {};
+	}
+
+	if(timeout < 0) {
+		co_await async::race_and_cancel(
+			async::lambda([&](auto c) -> async::result<void> {
+				co_await async::suspend_indefinitely(c, cancelEvent);
+				// if the cancelEvent was raised, we consider this wait to have been
+				// interrupted.
+				if (async::cancellation_token{cancelEvent}.is_cancellation_requested())
+					interrupted = true;
+			}),
+			async::lambda([&](auto c) -> async::result<void> {
+				if (req.has_signal_seq() && self->enteredSignalSeq() != req.signal_seq()) {
+					// a signal was already raised since the request's
+					// signal seqnum
+					interrupted = true;
+					co_return;
+				}
+				co_await async::suspend_indefinitely(c);
+			}),
+			async::lambda([&](auto c) -> async::result<void> {
+				k = co_await epoll::wait(epfile.get(), events, 16, c);
+			})
+		);
+	}else if(!timeout) {
+		// Do not bother to set up a timer for zero timeouts.
+		async::cancellation_event cancel_wait;
+		cancel_wait.cancel();
+		k = co_await epoll::wait(epfile.get(), events, 16, cancel_wait);
+	}else{
+		assert(timeout > 0);
+		co_await async::race_and_cancel(
+			async::lambda([&](auto c) -> async::result<void> {
+				// if the timeout runs to completion, i.e. the sleep does not return
+				// false to signal cancellation, we DO NOT consider the call to have
+				// been interrupted.
+				co_await helix::sleepFor(static_cast<uint64_t>(timeout), c);
+			}),
+			async::lambda([&](auto c) -> async::result<void> {
+				co_await async::suspend_indefinitely(c, cancelEvent);
+				// if the cancelEvent was raised, we consider this wait to have been
+				// interrupted.
+				if (async::cancellation_token{cancelEvent}.is_cancellation_requested())
+					interrupted = true;
+			}),
+			async::lambda([&](auto c) -> async::result<void> {
+				if (req.has_signal_seq() && self->enteredSignalSeq() != req.signal_seq()) {
+					// a signal was already raised since the request's
+					// signal seqnum
+					interrupted = true;
+					co_return;
+				}
+				co_await async::suspend_indefinitely(c);
+			}),
+			async::lambda([&](auto c) -> async::result<void> {
+				k = co_await epoll::wait(epfile.get(), events, 16, c);
+			})
+		);
+	}
+
+	// Assigned the returned events to each FD.
+	for(size_t j = 0; j < k; ++j) {
+		auto it = fdsToEvents.find(events[j].data.fd);
+		assert(it != fdsToEvents.end());
+
+		// Translate EPOLL events back to POLL events.
+		assert(!it->second.revents);
+		if(events[j].events & EPOLLIN) it->second.revents |= POLLIN;
+		if(events[j].events & EPOLLOUT) it->second.revents |= POLLOUT;
+		if(events[j].events & EPOLLPRI) it->second.revents |= POLLPRI;
+		if(events[j].events & EPOLLRDHUP) it->second.revents |= POLLRDHUP;
+		if(events[j].events & EPOLLERR) it->second.revents |= POLLERR;
+		if(events[j].events & EPOLLHUP) it->second.revents |= POLLHUP;
+	}
+
+	managarm::posix::EpollCallResponse resp;
+	bool hasEvents = false;
+
+	for(size_t i = 0; i < req.fds_size(); ++i) {
+		auto it = fdsToEvents.find(req.fds(i));
+		assert(it != fdsToEvents.end());
+		auto maskedEvents = it->second.revents & (req.events(i) | POLLHUP | POLLERR | POLLNVAL);
+		resp.add_events(maskedEvents);
+		if (!hasEvents && maskedEvents)
+			hasEvents = true;
+	}
+
+	if (!hasEvents && interrupted) {
+		resp.set_error(managarm::posix::Errors::INTERRUPTED);
+	} else {
+		resp.set_error(managarm::posix::Errors::SUCCESS);
+	}
+
+	auto [send_resp, send_tail] = co_await helix_ng::exchangeMsgs(conversation,
+		helix_ng::sendBragiHeadTail(resp, frg::stl_allocator{})
+	);
+	HEL_CHECK(send_resp.error());
+	HEL_CHECK(send_tail.error());
+	logBragiReply(resp);
+
 	co_return {};
 }
